requirements.txt
nltk==3.5
tweepy==3.10.0
tensorflow==2.4.1
numpy==1.19.5
boto3==1.15.1
python-dotenv==0.19.0
